// interfaces/lcu-types.d.ts
export interface LolStoreItemCost {
	currency: string
	/** @format int64 */
	cost: number
	/** @format float */
	discount?: number
}

export interface LolStoreItemKey {
	inventoryType: string
	/** @format int32 */
	itemId: number
}

export interface LolStoreItemLocalization {
	language: string
	name: string
	description: string
}

export interface LolStoreItemMetadataEntry {
	type: string
	value: string
}

export interface LolStoreItemOrderDTO {
	inventoryType: string
	/** @format int32 */
	itemId: number
	/** @format uint32 */
	quantity: number
	/** @format uint32 */
	rpCost: number
}

export interface LolStoreItemSale {
	/** @format uint64 */
	id: number
	active: boolean
	item: LolStoreItemKey
	sale: LolStoreSale
}
// ...
export interface LolStoreSale {
	startDate: string
	endDate: string
	prices: LolStoreItemCost[]
}